data_reader/requirements.txt
PyYAML==5.4.1
tabulate==0.8.7
